more test refactor
